feat: CookingInfo 컴포넌트 분할
